Day 2 part 1: Improve clarity with namedtuple
I wanted a way to label the two elements of the tuples in the
winning_games variable, so that you wouldn't have to work out whether
rock beats paper, or paper beats rock.

I'm not 100% happy with this approach; it's a lot of code just to to
identify that the first element is the opponent's shape, the second is
the current player's shape. A comment would have been just as good.