Day 7 part 1: List directories smaller than 100kB
This implementation feels reasonable, but the `from_log()` method would
bear a bit of refactoring.

In order to be able to parse the output of commands other than `ls` I'm
thinking it might make sense to introduce the concept of a current
command, and behaviour (e.g. a function) for parsing each line of its
output.

Given that part 1 hasn't needed that functionality, I'm not making that
change now (we should avoid hasty abstractions!)